requirements.txt
Flask
Flask-Alembic
Flask-JWT-Extended==3.24.1
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
Flask-Login
openpyxl
Jinja2==3.0.3
SQLAlchemy
Werkzeug==2.0.0
flask_script
PyMysqlDB
mysqlclient
apscheduler
requests
opencv-python
pika
python-daemon
python-dateutil
pyopenssl
mysql-connector-python
pymysql